Add endpoint and network configuration
This commit brings in endpoint and network configuration support.  This
support builds an appropriate struct sockaddr_{in,in6,un} from the
configuration file.  An "endpoint" is an IP address/port number pair (or
a local socket path) on which Humboldt listens, while a "network" is an
optional IP address/port number from which Humboldt may initiate a
connection.  Endpoints are "advertised" to the overlay network, and
the "networks" feature allows the administrator to configure which remote
endpoints a given Humboldt instance may connect to; this enables sane
handling of isolated domains, such as Humboldt instances behind NATs.