Create default endpoints
This change adds the support for creating default endpoints, if none are
read from the configuration.  This change also updates the behavior for
endpoints that don't have an 'ip' key specified (they will open ports on
all interfaces), and fixes a minor bug in ep_addr_default() related to
port handling in the case no 'ip' key was specified.