Update error handling for missing configuration files
This change updates the configuration file error handling to emit an error
if the configuration file could not be opened.  The only error that is
suppressed is if the configuration file is the default configuration file
and it is not found; this is handled sanely by the defaults.  If the
configuration file cannot be read for any reason, including YAML errors,
Humboldt exits.