Properly handle YAML implicit nodes
By itself, libyaml doesn't detect implicit nodes as anything other than
string scalars; that is, unless there's a "!!bool" tag on a node, a bare
"true" will appear to be a string even though the YAML standard would
call it a boolean.  Previously, this was handled by allowing bare strings
to substitute for the expected node, but this also was problemmatic, as
for instance integer conversion is much more rich than can be expressed
with, say, strtol().

This change introduces yaml_get_scalar(), which includes as part of its
called functions a hand-coded scanner which understands all of the
YAML 1.1 implicit nodes, including boolean, float (which we don't
convert), integer, null, string, and timestamp (which we also don't
convert).  The scanner implements the full range of recognition for these
nodes, so, e.g., "19_67" is properly recognized as a YAML integer.

As part of these changes, the yaml_get_str() function is updated to return
the length of the strings it returns, as well as grows rudimentary support
for binary nodes by changing the allow_null parameter into a flags
parameter.  Note that the actual binary decoding, which is base64, is not
yet supported.