Carefully handle setting IP addresses.
Prior to actually converting an IP address in ep_addr_set_ipaddr(), make
sure to re-clear the address structure.  This could be necessary if the
port is specified before the IP address and if the host places the port
number in different locations in struct sockaddr_in vs.
struct sockaddr_in6.