Use dynamic allocation for endpoints
Previously, we were using a flexlist_t to track endpoints allocated for
the runtime.  This resulted in errors, because the addresses of those
endpoints was being altered if the list needed to be reallocated.  This
change switches to using dynamic allocation through free lists for
allocating the endpoint structures, and to using a linked list to track
the endpoints in the runtime.