Create list_count() macro
I frequently make use of similar idioms to determine the number of entries
in a statically allocated list, such as the list of recognized facilities
in log.c.  This commit introduces the list_count() macro to common.h to
perform this calculation no matter the type of the elements of the list.