Implement deferred close
This change implements a deferred closing mechanism for connections that
ensures that data still in the buffer will be sent to the remote end.
This fixes the problem of the connection error packet not being received,
as well as preparing the way for SSL functionality: using the
CLOSE_ON_FREE bufferevent flag results in unclean SSL shutdown that
could interfere with the SSL session cache.