Implement basic STARTTLS message processing
This change adds the STARTTLS protocol (protocol 2).  This change really
starts to cause problems with recursive includes and starts to effect my
concepts for dealing with that problem, namely, putting all externally
referenced types before the internal includes.

Note that this change does *not* initiate TLS; it simply adds the
functions responsible for processing the STARTTLS protocol messages.