Add protocol 2 support to hum_proto
This change adds support for the STARTTLS protocol to the Humboldt
Protocol Analyzer.  It also adds support for default carrier flags, for
the benefit of the interface; without this, attempting to "send start tls
error" will actually end up sending a STARTTLS request.  This change also
explicitly sets the PROTOCOL attribute on the message classes, for the
same reason: attempting to "send start tls error" would actually end up
sending a RequestConnectionState message.