Rework receiver loop
This change reworks the receiver loop using queues to ensure that we can
wrap the socket without worrying about something currently trying to
receive on the socket.  This also makes use of the
prompt_toolkit.interface.CommandLineInterface.invalidate() method to
ensure that messages are displayed as soon as they are inserted into the
display buffer.