Vastly simplify ApplicationLoop
This change eliminates the use of threading entirely, as it was introduced
because of the problem that the discovery and use of the
prompt_toolkit.interface.CommandLineInterface.invalidate() happens to
eliminate, namely, the delay before the results of a call to
ApplicationLoop.display() are actually seen on screen.  This change also
introduces the commands "close" and "connect", for closing and
reconnecting by command; also, command aliases are introduced, so the
"quit" command is now an alias for the "exit" command.