Create an SSLContext
This change adds options for an SSL certificate and corresponding key
file.  This information is used to create an SSLContext (or, for older
versions of Python 2.7 which don't have that class, an instance of our
StubSSLContext), which is then passed to the ApplicationLoop.

With this change, the connection address is also made optional, allowing
connection to be specified using the "connect" command from within the
tool.