Simplify ApplicationLoop.wrap()
This change simplifies the ApplicationLoop.wrap() method by eliminating
the call to ApplicationLoop.display()--leaving that up to the caller--and
by raising reasonable exceptions in the case of errors.  That is, if the
socket is not open, a synthetic socket.error is raised with its errno
attribute set to EBADF, and any errors raised by the wrapper are allowed
to bubble up to the caller of ApplicationLoop.wrap().