Burn flexlist with fire
The flexlist implementation resulted in all sorts of problems, and was a
bad idea from the beginning, particularly for objects which were
referenced by other objects (e.g., endpoint configuration referenced by
endpoint advertisements).  With the addition of hash tables and the change
in how endpoint configuration is allocated, we can do away with flexlist
entirely.