Initialize SASL connection context
This change continues the SASL integration by creating a connection
context that can later be used for SASL exchanges.  This includes setting
a connection security strength factor and an external authentication ID,
which modify the mechanisms that SASL may advertise to a client.  Default
minimum security strength factors and the security strength factor
assigned to an endpoint (e.g., a local socket) can all be set from the
configuration file.  This is all properly integrated with TLS by setting
the connection's security strength factor for the connection from the bit
strength of the cipher algorithm.