Fix pbp_remaining(), add pbp_data(), enhance protocol_buf_append()
This change fixes a couple of bugs: 1. pbp_remaining() used the wrong
calculation (and was missing a doc comment); and
2. protocol_buf_append() should take a const char * instead of an
unsigned char *.  This change also makes two enhancements: 1. The
addition of pbp_data(), to obtain a pointer to the remaining data in a
protocol buffer; and 2. allow a size of -1 to protocol_buf_append() to
signal using the strlen() function on the data to obtain it.