Collect authenticated username
This change obtains the authenticated username from SASL after
the exchange completes.  This ensures that Humboldt will use the
proper username in subsequent authorization checks.  This change
also introduces a stub for creating security layers (right now,
just issues a warning if a security layer was negotiated) and
corrects a minor bug for clearing the SASL_INPROGRESS flag on the
connection.