Add connection mode
This change adds the concept of a connection mode, as well as
adding the remote connection state.  The connection_create()
function is updated to allow its endpoint argument to be
specified as NULL, which is used to select the initial connection
mode and type.  The initial connection state message is only sent
if the connection is in ACCEPTING mode.  These changes together
allow for a connection to be initiated by Humboldt, as well as
allowing for the turn-about that occurs after the connection
reaches the AUTH status.