State analysis
The next major phase of Humboldt is building up the state machine so that
Humboldt instances are able to fully authenticate to each other and secure
their communications.  The state description is complicated, however, so
it would be helpful to have some analysis tools to help visualize the
proposed state machines and perhaps perform other analysis.  This commit
contains a start on the state description and the analysis tools.  It
includes an "idea" documentation file regarding the states; a YAML file
describing the initial set of states and their transitions; a file in the
GraphViz "dot" format; a PNG file with a visualization of the state
transitions, generated by GraphViz from the "dot" file; and a tool capable
of creating the "dot" file from the YAML file, as well as of
canonicalizing the YAML file.  Future commits will include the actual
analysis tools, as they are created.